Major restructuring before implementing decode.
