+++ ethane-abi/src/lib.rs
@@ -9,8 +9,10 @@ use thiserror::Error;
 mod function;
 mod parameter;
 
-pub use function::{Function, StateMutability};
-pub use parameter::{Parameter, ParameterType};
+use function::Function;
+pub use function::StateMutability;
+pub use parameter::Parameter;
+use parameter::ParameterType;
 
 /// Parses a `.json` file containing ABI encoded Solidity functions.
 ///
@@ -64,68 +66,18 @@ impl Abi {
         self.parse_json(abi)
     }
 
-    pub fn get_state_mutability(&self, abi_call: &AbiCall) -> Option<StateMutability> {
-        if let Some(function) = self.functions.get(abi_call.function_name) {
+    pub fn get_state_mutability(&self, function_name: &str) -> Option<StateMutability> {
+        if let Some(function) = self.functions.get(function_name) {
             return function.state_mutability;
         }
 
         None
     }
 
-    /// Encodes a given [`AbiCall`] into a vector of bytes.
-    ///
-    /// The result's first 4 bytes are generated via a `Keccak256` hash of the
-    /// function signature. Check [Solidity's
-    /// documentation](https://docs.soliditylang.org/en/v0.5.3/abi-spec.html)
-    /// for more info.
-    pub fn encode(&self, abi_call: &AbiCall) -> Result<Vec<u8>, AbiParserError> {
-        if let Some(function) = self.functions.get(abi_call.function_name) {
-            // Check whether the correct number of parameters were provided
-            if abi_call.parameters.len() != function.inputs.len() {
-                return Err(AbiParserError::MissingData(format!(
-                    "Wrong number of parameters were provided. Expected {}, found {}",
-                    function.inputs.len(),
-                    abi_call.parameters.len(),
-                )));
-            }
-
-            // Create function signature and hash
-            let input_type_str = function
-                .inputs
-                .iter()
-                .map(|input| input.parameter_type.as_abi_string())
-                .collect::<Vec<String>>()
-                .join(",");
-
-            let signature = format!("{}({})", abi_call.function_name, input_type_str);
-            let mut keccak = Keccak256::new();
-            keccak.update(signature);
-            // Take first 4 bytes of the Keccak hash
-            let mut hash = keccak.finalize()[0..4].to_vec();
-            // Append the encoded parameters to the hash
-            for (parameter, input) in abi_call.parameters.iter().zip(function.inputs.iter()) {
-                if parameter.get_type() != input.parameter_type {
-                    return Err(AbiParserError::InvalidAbiEncoding(format!(
-                        "Invalid parameter type supplied. Expected {:?}, found {:?}",
-                        input.parameter_type,
-                        parameter.get_type()
-                    )));
-                }
-                hash.append(&mut parameter.encode());
-            }
-
-            Ok(hash)
-        } else {
-            Err(AbiParserError::MissingData(
-                "Function name not found in ABI".to_owned(),
-            ))
-        }
-    }
-
-    pub fn updated_encode(
+    pub fn encode(
         &self,
         function_name: &str,
-        parameters: Vec<parameter::tmp::Parameter>,
+        parameters: Vec<Parameter>,
     ) -> Result<Vec<u8>, AbiParserError> {
         if let Some(function) = self.functions.get(function_name) {
             let mut abi_arguments = Vec::<String>::with_capacity(parameters.len());
@@ -145,7 +97,7 @@ impl Abi {
             // Take first 4 bytes of the Keccak hash
             let mut hash = hasher.finalize()[0..4].to_vec();
             // Append the encoded parameters to the hash
-            parameter::encode_into::encode_into(&mut hash, parameters);
+            parameter::encode_into(&mut hash, parameters);
             Ok(hash)
         } else {
             Err(AbiParserError::MissingData(
@@ -182,12 +134,6 @@ impl Abi {
     }
 }
 
-/// An ABI call containing the called function's name and its input parameters.
-pub struct AbiCall<'a> {
-    pub function_name: &'a str,
-    pub parameters: Vec<Parameter>,
-}
-
 #[derive(Error, Debug)]
 pub enum AbiParserError {
     #[error("Couldn't open ABI file: {0}")]
@@ -202,6 +148,7 @@ pub enum AbiParserError {
     TypeError,
 }
 
+/*
 #[cfg(test)]
 mod tests {
     use std::str::FromStr;
@@ -284,3 +231,4 @@ mod tests {
         assert_eq!(hash.unwrap(), expected);
     }
 }
+*/
